Adding 16K ROM support to FEED low-level driver
Support is achieved by attempting to parse the usual 2K ROM and falling
back to the 16K ROM at the new starting location (0x4000) if that fails.

The driver architecture requires reading the full ROM aperture even if
it is not fully populated. Stick with this concept for now, although it
means that we're always reading 2K + 16K at startup.